+++ KingsMen/src/main/java/org/example/model/OrderDetails.java
@@ -17,7 +17,7 @@ public class OrderDetails {
     private String order_products;
     
     private Integer status;
-    private Long total;
+    private Double total;
 
     public Integer getId() {
         return id;
@@ -51,11 +51,11 @@ public class OrderDetails {
         this.status = status;
     }
 
-    public Long getTotal() {
+    public Double getTotal() {
         return total;
     }
 
-    public void setTotal(Long total) {
+    public void setTotal(Double total) {
         this.total = total;
     }
